Add tests of Magick class
